requirements.txt
Django>=4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
django-filter==23.3
drf-yasg==1.21.7
python-decouple==3.8
django-cors-headers==4.3.0
gunicorn==20.1.0

# Static Files (NEW - Required for production)
whitenoise==6.6.0

# Optional but recommended
Pillow  # For image handling if needed
